// src/app/features/unreal-pnl-query/models/unreal-pnl-query-request.model.ts
export interface UnrealPnlQueryRequest {
  APISERVER: string;
  sid: string;
  sip: string;
  Invscode: string;
  comp: string;
  bhno: string;
  cseq: string;
  stock: string;
  stockName: string;
  ttype: string;
  action: string;
  action1: string;
  action2: string;
  action4: string;
  action8: string;
  ctype: string;
  exchangeRate: string;
}